+++ src/test/java/com/transferwise/openbanking/client/oauth/ClientSecretBasicAuthenticationTest.java
@@ -2,7 +2,6 @@ package com.transferwise.openbanking.client.oauth;
 
 import com.transferwise.openbanking.client.configuration.AspspDetails;
 import com.transferwise.openbanking.client.oauth.domain.GetAccessTokenRequest;
-import com.transferwise.openbanking.client.test.TestAspspDetails;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -12,6 +11,8 @@ import org.springframework.util.Base64Utils;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.List;
+
+import static com.transferwise.openbanking.client.test.factory.AspspDetailsFactory.aAspspDetails;
 
 class ClientSecretBasicAuthenticationTest {
 
@@ -31,7 +32,7 @@ class ClientSecretBasicAuthenticationTest {
     @Test
     void addClientAuthentication() {
         GetAccessTokenRequest getAccessTokenRequest = GetAccessTokenRequest.clientCredentialsRequest("payments");
-        AspspDetails aspspDetails = aAspspDefinition();
+        AspspDetails aspspDetails = aAspspDetails();
 
         String expectedBearerToken = Base64Utils.encodeToString(
             (aspspDetails.getClientId() + ":" + aspspDetails.getClientSecret())
@@ -42,12 +43,5 @@ class ClientSecretBasicAuthenticationTest {
 
         List<String> authorisationHeaders = getAccessTokenRequest.getRequestHeaders().get(HttpHeaders.AUTHORIZATION);
         Assertions.assertEquals(Collections.singletonList(expectedAuthorisation), authorisationHeaders);
-    }
-
-    private AspspDetails aAspspDefinition() {
-        return TestAspspDetails.builder()
-            .clientId("client-id")
-            .clientSecret("client-secret")
-            .build();
     }
 }