+++ src/test/java/com/transferwise/openbanking/client/oauth/ClientSecretPostAuthenticationTest.java
@@ -2,10 +2,11 @@ package com.transferwise.openbanking.client.oauth;
 
 import com.transferwise.openbanking.client.configuration.AspspDetails;
 import com.transferwise.openbanking.client.oauth.domain.GetAccessTokenRequest;
-import com.transferwise.openbanking.client.test.TestAspspDetails;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import static com.transferwise.openbanking.client.test.factory.AspspDetailsFactory.aAspspDetails;
 
 class ClientSecretPostAuthenticationTest {
 
@@ -25,19 +26,12 @@ class ClientSecretPostAuthenticationTest {
     @Test
     void addClientAuthentication() {
         GetAccessTokenRequest getAccessTokenRequest = GetAccessTokenRequest.clientCredentialsRequest("payments");
-        AspspDetails aspspDetails = aAspspDefinition();
+        AspspDetails aspspDetails = aAspspDetails();
 
         clientSecretPostAuthentication.addClientAuthentication(getAccessTokenRequest, aspspDetails);
 
         Assertions.assertEquals(aspspDetails.getClientId(), getAccessTokenRequest.getRequestBody().get("client_id"));
         Assertions.assertEquals(aspspDetails.getClientSecret(),
             getAccessTokenRequest.getRequestBody().get("client_secret"));
-    }
-
-    private AspspDetails aAspspDefinition() {
-        return TestAspspDetails.builder()
-            .clientId("client-id")
-            .clientSecret("client-secret")
-            .build();
     }
 }