+++ src/test/java/com/transferwise/openbanking/client/test/factory/AuthorizationContextFactory.java
@@ -1,0 +1,14 @@
+package com.transferwise.openbanking.client.test.factory;
+
+import com.transferwise.openbanking.client.api.common.AuthorizationContext;
+
+public class AuthorizationContextFactory {
+
+    public static AuthorizationContext aAuthorizationContext() {
+        return AuthorizationContext.builder()
+            .authorizationCode("authorisation-code")
+            .redirectUrl("https://tpp.co.uk")
+            .build();
+    }
+
+}